# backend/api/openapi.yaml
openapi: 3.1.1
info:
  title: Sama HTTP contract
  version: 0.0.0
  description: >-
    This source contract documents implemented liveness, OIDC login/logout,
    current-user and workspace routes and the shared HTTP problem envelope. Future entity identifiers use canonical
    lowercase hyphenated UUIDv7 strings; request_id is a separate bounded
    correlation value and is not an entity identifier or secret.
servers:
  - url: /
paths:
...
  /auth/login:
    get:
      operationId: startLogin
      summary: Start a bounded OIDC login with a one-use browser-bound challenge
      responses:
        '302':
          description: Redirect to the configured issuer with state, nonce and S256 PKCE.
          headers:
            Location:
              schema:
                type: string
            Set-Cookie:
              schema:
                type: string
        default:
          $ref: '#/components/responses/Problem'

    post:
      operationId: startInvitedLogin
      summary: Bind invitation proof to the browser challenge; admission waits for verified OIDC callback
      parameters:
        - {name: Origin, in: header, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [invitation]
              properties:
                invitation: {type: string, minLength: 43, maxLength: 43}
      responses:
        '302':
          description: Browser-bound OIDC redirect. This preauthentication initiation requires configured Origin and JSON; it creates no session or membership.
        default:
          $ref: '#/components/responses/Problem'
  /auth/reauthenticate:
    post:
      operationId: reauthenticate
      security: [{session: []}]
      parameters:
        - {name: Origin, in: header, required: true, schema: {type: string}}
        - {name: X-CSRF-Token, in: header, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, additionalProperties: false}
      responses:
        '302':
          description: Session-bound OIDC redirect using prompt=login, max_age=0 and configured ACR. Callback requires fresh auth_time and rotates the same identity's session.
        default:
          $ref: '#/components/responses/Problem'
  /auth/callback:
    get:
      operationId: finishLogin
      summary: Consume a browser-bound OIDC challenge and issue an admitted user session
      parameters:
        - {name: state, in: query, required: true, schema: {type: string}}
        - {name: code, in: query, required: true, schema: {type: string}}
      responses:
        '302':
          description: Session cookie issued; redirect to the same-origin overview.
          headers:
            Set-Cookie:
              schema: {type: string}
        default:
          $ref: '#/components/responses/Problem'
  /auth/logout:
    post:
      operationId: logout
      summary: Revoke the current session and expire its cookie
      parameters:
        - {name: Origin, in: header, required: true, schema: {type: string}}
        - {name: X-CSRF-Token, in: header, required: true, schema: {type: string}}
      responses:
        '204':
          description: Session revoked or already absent; session cookie expired.
          headers:
            Set-Cookie:
              schema: {type: string}
        default:
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    session: {type: apiKey, in: cookie, name: __Host-sama_session}
...
  headers:
    RequestId:
      description: Bounded request correlation identifier.
      schema:
        type: string
        minLength: 1
        maxLength: 64
        pattern: '^[A-Za-z0-9._-]+$'
  responses:
    Problem:
      description: Stable HTTP error envelope. Unknown routes do not imply domain operations.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
        Cache-Control:
          schema:
            type: string
            const: no-store
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    Problem:
      type: object
      additionalProperties: false
      required:
        - type
        - title
        - status
        - code
        - request_id
      properties:
        type:
          type: string
          const: about:blank
        title:
          type: string
        status:
          type: integer
          enum:
            - 408
            - 413
            - 431
            - 409
            - 422
            - 403
            - 401
            - 400
            - 404
            - 405
            - 429
            - 503
        code:
          type: string
          enum:
            - reauthentication_unsupported
            - recent_authentication_required
            - body_too_large
            - headers_too_large
            - request_cancelled
            - policy_conflict
            - membership_conflict
            - not_found
            - method_not_allowed
            - invalid_workspace
            - workspace_limit
            - permission_denied
            - unauthenticated
            - invalid_request
            - rate_limited
            - identity_unavailable
        request_id:
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^[A-Za-z0-9._-]+$'